requirements.txt
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
pytz==2024.1
